+++ src/app/shared/components/error-back-text/error-back-text.component.ts
@@ -1,4 +1,5 @@
 import { Component, Input, OnInit } from '@angular/core';
+import { StaticService } from 'src/app/config/static.service';
 
 @Component({
   selector: 'fs-error-back-text',
@@ -7,7 +8,10 @@ import { Component, Input, OnInit } from '@angular/core';
 export class ErrorBackTextComponent implements OnInit {
   @Input() errorText: string;
 
-  constructor() {}
+  lnkLanding = this.staticConfig.getPathInfo().lnkLanding;
+  lnkOffers = this.staticConfig.getPathInfo().lnkOffers;
+
+  constructor(private staticConfig: StaticService) {}
 
   ngOnInit(): void {
   }