+++ src/app/sites/offers/components/modalWindows/modal-offer-to-lms/ngbd-modal-lmscontent.ts
@@ -1,0 +1,13 @@
+import { Component, Input } from '@angular/core';
+import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
+
+@Component({
+  selector: 'ngbd-modal-lms-content',
+  templateUrl: './ngbd-modal-lmscontent.html',
+  styleUrls: ['./ngbd-modal-lmscontent.scss'],
+})
+export class NgbdModalLmsContent {
+  @Input() title;
+
+  constructor(public activeModal: NgbActiveModal) {}
+}