+++ Changelog.md
@@ -11,3 +11,4 @@ New Project with Angular 11
 - New Testframework Jest
 - Folder Structure
 - Icon Library only contains used Icons (not the complete library)
+- All Links are now variables. See static.config.ts